Deprecated Embed types
Closes #80

Discord has now officially documented the existing embed types, and consequently deprecated them as well.

This commit adds the embed types we were still missing and deprecates the type enum as well as the type field in embeds. There doesn't seem to be any replacement for this field, as such none is offered in our deprecation either.

Lastly a small typo was fixed: embets -> embeds.