Add preferredLocale and publicUpdatesChannelId
Closes #81

Turns out we were missing preferred locale as well on both guild and guild modifies, so I've taken the liberty to add those as well. The documentation doesn't show those to be present on guild create though.

Added a small test to the modify builder as well to make sure Locale is correctly converted, Discord's documentation leads me to believe that they only accept language and region tags.