Include json error in exception message (#369)
* Include json error in exception message

Changed the `RestRequestException` message to include the JSON error message. `DiscordFieldError` was removed, it made it hard to reconstruct the message and is not documented by Discord. It was replaced by the literal json.

This does mean that rest exceptions are no longer serialization agnostic (relying on kx.ser-json), the alternative could be to deser the field as a literal JSON string.

Additionally, the request that caused the error was added to the exception, but omitted from the message for brevity.

* Add DiscordErrorResponse defaults