Fix deprecation propagation for new select menus
The deprecations for SelectMenuBuilder.options and
SelectMenuComponent.options were propagated to the overrides in
StringSelectBuilder and StringSelectComponent when a user of Kord didn't
have progressive mode enabled.

To fix this, extensions are provided instead.

The changes in the public API are binary incompatible compared to #707
but not compared to 0.8.0-M17 so only snapshot users can be affected.

see #707, https://youtrack.jetbrains.com/issue/KT-47902 and
https://discord.com/channels/556525343595298817/587324906702766226/1042799716745879643