Fix forum incompatibilities in core
NewsChannelBehavior was missing TopGuildMessageChannelBehavior as a
supertype since ThreadParentChannelBehavior is no longer a
TopGuildMessageChannelBehavior.

TextChannelBehavior was missing TopGuildMessageChannelBehavior as a
supertype since PrivateThreadParentChannelBehavior is no longer a
TopGuildMessageChannelBehavior.

ThreadParentChannelBehavior was extended to be a
CategorizableChannelBehavior (implicating that ThreadParentChannel was
als extended to be a CategorizableChannel)

TextChannelBehavior's startPublicThread and startPrivateThread had an
additional reason parameter added, breaking binary compatibility. The
reason can be specified in the builder lambda.

TextChannelBehavior.startPublicThreadWithMessage had the archiveDuration
parameter removed, breaking binary compatibility.