Add inspection for missing Optional default values
If a dev.kord.common.entity.optional.Optional is used in a
@Serializable class, it should always have a default value so that
deserialization works if the value is missing in the JSON.

Until now this was easy to forget which could lead to deserialization
errors down the line.

By introducing a KSP processor that looks at all primary constructor
parameters of @Serializable classes, there is now a static check
executed at build time that verifies the default value is not forgotten
in any place.