Replace Icon with Asset (#703)
This introduces the new dev.kord.core.entity.Asset class that is a
replacement for dev.kord.core.entity.Icon with coverage of more CDN
endpoints and fewer inconsistencies (e.g. format was not considered when
calling Icon.getImage() with no parameters).

Properties with type Icon are now also deprecated and are replaced with
similar properties with type Asset. This means there are a few source-
(but not binary-) incompatible changes:

 * The types of dev.kord.core.entity.GuildEmoji.image,
   dev.kord.core.entity.Member.memberAvatar,
   dev.kord.core.entity.Role.icon, dev.kord.core.entity.User.avatar and
   dev.kord.core.entity.User.defaultAvatar were changed from Icon to
   Asset.

 * The type of dev.kord.core.entity.Team.icon was changed from String to
   Asset. Team.iconHash provides the previous functionality.

---------

Co-authored-by: Lukellmann <lukellmann@gmail.com>