Update dependencies (#826)
* Kotlin 1.8.10 -> 1.8.21
* Ktor 2.2.4 -> 2.3.0
* kotlinx.coroutines 1.6.4 -> 1.7.1
* kotlinx.serialization 1.5.0 -> 1.5.1
* kotlin-node 18.14.0-pre.502 -> 18.16.3-pre.546
* KSP 1.8.10-1.0.9 -> 1.8.21-1.0.11
* KotlinPoet 1.12.0 -> 1.13.2
* JUnit 5 5.9.2 -> 5.9.3
* MockK 1.13.4 -> 1.13.5
* AtomicFU 0.20.0 -> 0.20.2
* Binary compatibility validator 0.13.0 -> 0.13.1
* gradle-buildconfig-plugin 3.1.0 -> 4.0.4

Warnings about duplicate library name with AtomicFU are no longer
present with kotlinx.coroutines 1.7.1 and AtomicFU 0.20.2, so -Werror
can be turned on again.

Using typealiases for overriding the default Instant serializer wasn't
possible in #605 because of
https://github.com/Kotlin/kotlinx.serialization/issues/1895.
Since this Bug was fixed in Kotlin 1.8.21, the typealiases
dev.kord.common.serialization.InstantInEpochMilliseconds and
dev.kord.common.serialization.InstantInEpochSeconds can now be used.

The changes in the public API are probably because of changes in the
serialization and AtomicFU plugins with Kotlin 1.8.21. However, they
are unlikely to affect anyone since only generated symbols that aren't
supposed to be accessed directly were touched.