Don't use value class Reset with AtomicRef (#863)
The use of value classes for the Reset class has already caused issues
implementing #855. However, it was isolated to native platforms.
Unfortunately, this behavior will become a compiler error soon, so we
need to change it. This can be achieved by simply using the backing
Instant instead of Reset with AtomicRef.

Related Issues:
#855
#69
Kotlin/kotlinx-atomicfu#291
https://youtrack.jetbrains.com/issue/KT-61584

Co-authored-by: lukellmann <lukellmann@gmail.com>