Fix typo in AtomicFU property
I added this property in #894 to ensure we'd be using the AtomicFU
compiler plugin [1] for native when #855 is merged. However, it seems
like a typo slipped through: I think I copied the property from the
description of the PR that added this feature to AtomicFU [2]. Turns out
there was a typo in that description. The property is
kotlinx.atomicfu.enableNativeIrTransformation, not
kotlinx.atomicfu.enableNativeIrTransformations as can be seen in the
source code [3].

[1] https://github.com/Kotlin/kotlinx-atomicfu/blob/0.23.1/README.md#atomicfu-compiler-plugin
[2] https://github.com/Kotlin/kotlinx-atomicfu/pull/363
[3] https://github.com/Kotlin/kotlinx-atomicfu/blob/0.23.1/atomicfu-gradle-plugin/src/main/kotlin/kotlinx/atomicfu/plugin/gradle/AtomicFUGradlePlugin.kt#L38