Update Kotlin to 2.0.0 (#942)
The AutoCloseable interface is now stable and replaces the use of
io.ktor.utils.io.core.Closeable for dev.kord.gateway.Inflater.

KT-51045 [1] is fixed with K2, so a new type parameter is added to
BaseChoiceBuilder and NumericOptionBuilder.

The changes in the JVM ABI are caused by changes in the serialization
compiler plugin for K2 relative to the plugin for K1. They don't affect
binary compatibility.

The changes in the KLib ABI dumps (except for BaseChoiceBuilder and its
inheritors) are only syntactic changes with Kotlin 2.0.0 (e.g. [2]).

[1] https://youtrack.jetbrains.com/issue/KT-51045
[2] https://github.com/JetBrains/kotlin/commit/6972ed67d24644cd143861f21e398508d1689b3f

Co-authored-by: Luca Kellermann <lukellmann@gmail.com>